# docs/swagger.yaml
...
swagger: '2.0'
schemes:
  - http
host: localhost:5000
basePath: /
info:
  description: |
    # ThriftBox REST API
    ThriftBox is an application to manage and administer cooperative societies.
  version: '1.0.0'
  title: ThriftBox
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  x-logo:
    url: '/ale/logo.png'
...
produces:
  # This is a global default. You can OVERWRITE it in each specific operation.
  - application/json
consumes:
  # List of mime types the API endpoints consume
  # This is a global default. You can OVERWRITE it in each specific operation.
  - application/json
securityDefinitions:
  # Defines the security implemented in the application. The token security system is implemented.
  tokenAuth:
    type: basic
# Holds the relative paths to the individual endpoints. The path is appended to the basePath in order to construct the full URL.
paths:
...
  '/books/{bookId}/ledger':
    parameters:
      - name: bookId
        description: The book to extract entries from
        in: path
        type: integer
        required: true
    post:
      security:
        - tokenAuth: []
      summary: Post a new Journal entry
      description: Add a new (balanced) Journal entry to the book
      operationId: postBookEntry

      parameters:
        - name: body
          in: body
          description: The new Journal Entry definition
          schema:
            $ref: '#/definitions/NewJournalEntry'

      produces:
        - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            $ref: '#/definitions/PostEntryResponse'

    get:
      security:
        - tokenAuth: []
      summary: Fetch the ledger
      description: |
        Fetches all transactions for the given book for the dates provided
      operationId: getBookEntries

      parameters:
      - name: bookId
        description: The book to extract entries from
        in: path
        type: integer
        required: true

      - name: startDate
        in: query
        description: 'The start date for entries'
        required: false
        type: string

      - name: endDate
        in: query
        description: 'The end date for entries'
        required: false
        type: string

      - name: perPage
        in: query
        description: 'The number of results per page'
        required: false
        type: integer

      - name: page
        in: query
        description: 'The page number'
        required: false
        type: integer

      produces:
      - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            $ref: '#/definitions/Entries'
          examples:
            application/json:
              {
                book: { id: "General Ledger", currency: 'USD' },
                startDate: 1520228313023,
                endDate: 1520428313023,
                entries: [
                  {
                    date: 1520228313023,
                    memo: 'Payroll',
                    transactions: [
                      { account: "Bank:Local", debit: 1000.0, currency: "USD", exchangeRate: 1.0 },
                      { account: "Employees:Alice", value: 500.0, currency: "USD", exchangeRate: 1.0 },
                      { account: "Employees:Bob", value: 500.0, currency: "USD", exchangeRate: 1.0 }
                    ]
                  }
                ]
              }
        400:
          description: Invalid input, such as unknown book
          schema:
            $ref: '#/definitions/Response'
...
  '/books/{bookId}/transactions':
    parameters:
    - name: bookId
      description: The book to extract entries from
      in: path
      type: integer
      required: true
    get:
      security:
        - tokenAuth: []
      summary: List all transactions for given accounts
      operationId: getTransactions
      parameters:
      - name: account
        description: Name of account to return transactions for
        in: query
        type: string
        required: true

      - name: accountCode
        description: Code of account to return transactions for
        in: query
        type: number
        required: true

      - name: perPage
        in: query
        description: 'The number of results per page'
        required: false
        type: integer

      - name: page
        in: query
        description: 'The page number'
        required: false
        type: integer

      produces:
        - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Transaction'
  '/books/{bookId}/balance':
    parameters:
    - name: bookId
      description: The book to extract entries from
      in: path
      type: integer
      required: true
    get:
      security:
        - tokenAuth: []
      summary: Get account balance
      description: Returns an account balance
      operationId: getBalance
      parameters:
      - name: account
        description: The account name to get the balance for
        in: query
        type: string
        required: true
      - name: accountCode
        description: The account code to get the balance for
        in: query
        type: string
        required: true
      - name: inQuoteCurrency
        description: If true (default), converts all values to the quote currency first
        in: query
        type: boolean
        required: false
      produces:
        - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            $ref: '#/definitions/Balance'
...
  '/books/{bookId}/subaccttransactions':
    parameters:
    - name: bookId
      description: The book to extract entries from
      in: path
      type: integer
      required: true
    get:
      security:
        - tokenAuth: []
      summary: Fetch sub account transactions
      description: Return an array of sub account transactions
      operationId: getSubAcctTransactions
      parameters:
      - name: account
        description: The account name to get the transactions for
        in: query
        type: string
        required: true
      - name: accountCode
        description: The account code to get the transactions for
        in: query
        type: string
        required: true
      - name: userId
        description: The id of user who owns the subaccount
        in: query
        type: string
      - name: companyId
        description: The id of company who owns the subaccount
        in: query
        type: string
      - name: inQuoteCurrency
        description: If true (default), converts all values to the quote currency first
        in: query
        type: boolean
        required: false
      produces:
        - application/json
      responses: # list of responses
        200:
          description: Success
          schema:
            $ref: '#/definitions/SubTransactions'
...
definitions:
...
  Book:
    type: object
    required:
      - name
      - currency
    properties:
      id:
        description: The id for the book the book
        type: integer
      name:
        description: The name of the book
        type: string
        example: General Ledger
      currency:
        description: |
          The currency the book is referenced in.
          All other currencies and calculations are quoted in terms of this currency
        type: string
        example: USD
      createdAt:
        description: The timestamp of when the book was created
        type: number
      updatedAt:
        description: The timestamp of the last time this entry was modified
        type: number
  Entries:
    type: object
    properties:
      book:
        $ref: '#/definitions/Book'
      startDate:
        type: number
      endDate:
        type: number
      entries:
        type: array
        items:
          $ref: '#/definitions/Entry'
  Entry:
    type: object
    properties:
      date:
        description: The timestamp for the entry
        type: number
      memo:
        description: The description for the entry
        type: string
      voided:
        description: Indicates whether the entry is still valid
        type: boolean
      voidReason:
        description: The reason for the entry reversal
        type: string
      transactions:
        description: An array of transactions for the entry
        type: array
        items:
          $ref: '#/definitions/Transaction'
  Transaction:
    type: object
    required:
      - account
      - accountCode
    properties:
      account:
        description: The account name this transaction is reflected on
        type: string
      accountCode:
        description: The account code this transaction is reflected on
        type: number
      credit:
        description: The credit value of the transaction
        type: number
      debit:
        description: The debit value of the transaction
        type: number
      currency:
        description: The currency for this transaction
        type: string
      exchangeRate:
        description: The exchange rate to convert to the basis currency
        type: number
  Response:
    type: object
    properties:
      success:
        description: Indicates whether request was succesful
        type: boolean
      message:
        type: string
...
  PostEntryResponse:
    allOf:
      - $ref: '#/definitions/Response'
      - properties:
          id:
            description: The id of the new Journal Entry
            type: integer
  NewJournalEntry:
    type: object
    required:
      - memo
      - transactions
    properties:
      memo:
        description: The Journal Entry description
        type: string
      timestamp:
        description: The time stamp for the journal entry
        type: string
      transactions:
        type: array
        items:
          $ref: '#/definitions/Transaction'
...
  Balance:
    type: object
    properties:
      success:
        description: Indicates whether request was succesful
        type: boolean
      message:
        description: Short elaboration of status
        type: string
      data:
        description: An object containing the following properties - creditTotal, debitTotal, balance, current, number of transactions
        type: object
  SubTransactions:
    type: object
    properties:
      success:
        description: Indicates whether request was succesful
        type: boolean
      message:
        description: Short elaboration of status
        type: string
      data:
        description: An array of objects containing individual transaction entries 
        type: array
        items: {}
...